Added new waitFor helpers
